refactor: move firebase app initialization out of AuthService
Firebase should only be initialized once. Since multiple services
will be using firebase, it should be initialized outside of any
single service.
